==> Code/3b_ssurgo_process.R <==

library(tidyverse)
library(sf)
library(readxl)

# ==== LOAD ====================================================================

# Load fields panel
fields = st_read("Data/Clean/Fields/fields_panel.gpkg") |> 
  # Filter to 2024 fields
  filter(year == 2024) |> 
  # Transform to NAD83 for spatial operations
  st_transform(crs = 26912)

# Load map unit polygons
mu_polys = st_read("Data/Raw/SSURGO/ssurgo_ut.gpkg", layer = "mu_polys") |> 
  # Transform to NAD83 for spatial operations
  st_transform(crs = 26912)

# Load component tabular data
comp = st_read("Data/Raw/SSURGO/ssurgo_ut.gpkg", layer = "comp")

# Load component restrictions
co_restrictions = st_read("Data/Raw/SSURGO/ssurgo_ut.gpkg", layer = "co_restrictions")

# Load map unit aggregated attribute data
mu_aggatt = st_read("Data/Raw/SSURGO/ssurgo_ut.gpkg", layer = "mu_aggatt")

# Load curve numbers
curve_numbers = read_xlsx("Data/Misc/curve_numbers.xlsx")

# ==== PROCESS =================================================================

# Intersect SSURGO map units with area of interest
ssurgo_aoi = st_intersection(mu_polys, fields)

# Calculate component-weighted mean depth to restrictive layer for each map unit in inches
mu_restrictions = co_restrictions |> 
  # Join component keys with map unit keys and component percents
  left_join(comp, by = "cokey") |>
  # Calculate component-weighted mean depth to restrictive layer for each map unit in inches
  group_by(mukey) |> 
  summarize(
    restrictive_layer_in = weighted.mean(restrictive_layer_cm, w = co_pct, na.rm = FALSE) / 2.54,
    .groups = "drop"
  )

# Calculate mean curve number for each hydrologic group
cn_mean = curve_numbers |> 
  group_by(hydro_group) |> 
  summarize(mean_cn = mean(curve_number, na.rm = TRUE), .groups = "drop")

# Calculate mean curve number for multiple hydrologic groups
cn_mean = cn_mean |> 
  bind_rows(
    tibble(
      hydro_group = c("A/D", "B/D", "C/D"),
      mean_cn = c(
        mean(cn_mean$mean_cn[cn_mean$hydro_group %in% c("A","D")]),
        mean(cn_mean$mean_cn[cn_mean$hydro_group %in% c("B","D")]),
        mean(cn_mean$mean_cn[cn_mean$hydro_group %in% c("C","D")])
      )
    )
  )

# Calculate component percent-weighted mean curve number for each map unit
mu_cn = comp |>
  # Join with hydrologic group curve numbers
  left_join(cn_mean, by = "hydro_group", relationship = "many-to-one") |> 
  group_by(mukey) |> 
  summarize(
    curve_number = if_else(
      # If all of a map unit's components are missing CN, assign as NA
      all(is.na(mean_cn)),
      NA,
      # Compute weighted mean CN of non-NA components
      sum(mean_cn * co_pct, na.rm = TRUE) / sum(co_pct[!is.na(mean_cn)], na.rm = TRUE)
    ),
    .groups = "drop"
  )

ssurgo_stats = ssurgo_aoi |> 
  # Join map units with their AWS and depth to water table
  left_join(mu_aggatt, by = "mukey", relationship = "many-to-one") |> 
  # Join map units with their depth to restrictive layer
  left_join(mu_restrictions, by = "mukey", relationship = "many-to-one") |> 
  # Join map units with their hydrologic condition and group
  left_join(mu_cn, by = "mukey", relationship = "many-to-one") |> 
  mutate(
    # Compute area in acres of each map unit polygon
    mu_acres = as.numeric(st_area(geom)) / 4046.8564224,
    # Convert depth from cm to in
    water_table_in = water_table_cm / 2.54,
    # Convert available water storage from cm to in
    aws_in = aws_cm / 2.54
  )

ssurgo = ssurgo_stats |> 
  st_drop_geometry() |> 
  group_by(id) |> 
  summarize(
    # Calculate area-weighted mean AWS
    aws_in = weighted.mean(aws_in, w = mu_acres, na.rm = TRUE),
    # Calculate available water capacity
    awc_in_in = aws_in / 59,
    # Calculate area-weighted mean depth to water table
    water_table_in = weighted.mean(water_table_in, w = mu_acres, na.rm = TRUE),
    # Calculate area-weighted mean depth to restrictive layer
    restrictive_layer_in = weighted.mean(restrictive_layer_in, w = mu_acres, na.rm = TRUE),
    # Calculate area-weighted curve number
    curve_number = weighted.mean(curve_number, w = mu_acres, na.rm = TRUE),
    .groups = "drop"
  ) |> 
  mutate(
    # Max rooting depth possible based on water table or restrictive layer
    max_rz_in = pmin(water_table_in, restrictive_layer_in, na.rm = TRUE),
    # Convert NaN values to NA
    across(c(awc_in_in, max_rz_in), ~na_if(., NaN))
  ) |> 
  # Join each field with its SSURGO data
  right_join(fields, by = "id") |> 
  # Select needed variables
  select(id, awc_in_in, max_rz_in, curve_number)

# ==== SAVE ====================================================================

save(ssurgo, file = "Data/Clean/Input Data/ssurgo.rda")
